Migrate JSON and DAO to SDK
