+++ c3r-cli/build.gradle
@@ -78,8 +78,6 @@ dependencies {
     implementation project(":c3r-sdk-core")
     implementation project(":c3r-sdk-parquet")
 
-    implementation 'software.amazon.awssdk:cleanrooms:2.20.25'
-
     // Logging
     implementation 'org.apache.logging.log4j:log4j-api:2.20.0'
     implementation 'org.apache.logging.log4j:log4j-core:2.20.0'
