+++ c3r-sdk-core/src/main/java/com/amazonaws/c3r/json/ValidationTypeAdapterFactory.java
